+++ example/demo_india.dart
@@ -3,5 +3,5 @@ import 'package:tithi_engine/data/india.dart';
 
 void main() {
   final panchang = Panchang([registerIndia]);
-  print(panchang.forDate(DateTime(2026, 2, 15), 'Ujjain').displayName);
+  print(panchang.tithiOnDate(DateTime(2026, 2, 15), 'Ujjain').displayName);
 }