+++ lib/src/regions/bogota/boundary_corrections.dart
@@ -1,4 +1,0 @@
-const bogotaAmavasyaCorrections = <int, int>{};
-const bogotaPurnimaCorrections = <int, int>{
-  28334: 28333,
-};