+++ lib/src/regions/karachi-jni/boundary_corrections.dart
@@ -1,6 +1,0 @@
-const karachiAmavasyaCorrections = <int, int>{
-  62664: 62663,
-};
-const karachiPurnimaCorrections = <int, int>{
-  54529: 54528,
-};