+++ lib/src/regions/saopaulo/boundary_corrections.dart
@@ -1,5 +1,0 @@
-const saopauloAmavasyaCorrections = <int, int>{};
-const saopauloPurnimaCorrections = <int, int>{
-  57156: 57155,
-  67905: 67904,
-};